# month-and-math-service/src/main/java/com/company/monthandmathservice/models/CustomErrorResponse.java
package com.company.monthandmathservice.models;

import com.fasterxml.jackson.annotation.JsonFormat;

import java.time.LocalDateTime;

public class CustomErrorResponse {



            String errorMsg;
            int status;
            String errorCode;
            @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd hh:mm:ss")
            LocalDateTime timestamp;

            public CustomErrorResponse() {
            }

            public CustomErrorResponse(String errorCode, String errorMsg) {
                this.errorCode = errorCode;
                this.errorMsg = errorMsg;
            }

            public String getErrorCode() {
                return errorCode;
            }

            public void setErrorCode(String errorCode) {
                this.errorCode = errorCode;
            }

            public String getErrorMsg() {
                return errorMsg;
            }

            public void setErrorMsg(String errorMsg) {
                this.errorMsg = errorMsg;
            }

            public int getStatus() {
                return status;
            }

            public void setStatus(int status) {
                this.status = status;
            }
                public LocalDateTime getTimestamp() {
            return timestamp;
        }

        public void setTimestamp(LocalDateTime timestamp) {
            this.timestamp = timestamp;
        }
    }